Rename to opcode.rs
